+++ lib/index.js
@@ -3,7 +3,7 @@
 var Hapi = require('hapi');
 var Hoek = require('hoek');
 var Package = require('../package.json');
-
+var versionPlugin = require('./version');
 
 // Declare internals
 
@@ -15,16 +15,13 @@ internals.init = function () {
     var server = new Hapi.Server();
     server.connection({ port: 8000 });
 
-    server.route({
-        method: 'GET',
-        path: '/version',
-        config: {
-            description: 'Returns the version of the server',
-            handler: function (request, reply) {
+    server.register({register: versionPlugin }, {
+	routes: {
+            prefix: '/version'
+	}
+    }, function (err) {
 
-                return reply({ version: Package.version });
-            }
-        }
+        Hoek.assert(!err, err);
     });
 
     server.start(function (err) {
